Added max zoom to match the actual data I'm using.
I think this is partly responsible for the problem where waking the map from
sleep would only show the tile you were on before and not load any of the other
tiles if you were beyond the zoom level when there was actual data. Mapzen is
good about just showing data from a higher zoom level if it doesn't have
anything for the current zoom level, but apparently not after you wake from
sleep.