Move mapController out into the fragment so view model only manages state
This feels a bit cleaner and like the view model isn't doing stuff it shouldn't,
but it also isn't surviving rotation as well, maybe b/c the mapController is
getting trashed on rotation, while before it was surviving right up until it got
replaced by a new one. Maybe the mapController belongs in the view model, even
if it doesn't get used much.