Scene file selector; zoom futzing; ditched custom toolbar style
A note about the toolbar: in a perfect world I'd like the text, the hamburger
icon, and the contextual action icons all to have the same drop shadow effect,
but I can't seem to make that work for the vector drawables, so I'm just keeping
it simple with a vanilla toolbar. This commit deletes a lot of useful stuff
showing how to customize and manipulate the toolbar, so maybe worth referring
back to.